requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
sqlmodel==0.0.14
python-dotenv==1.0.0
pydantic>=2.12.3,<3
pydantic-settings>=2.5.2,<3
mcp>=1.11.0,<2
psycopg[binary]==3.3.1
python-jose[cryptography]==3.3.0
passlib[argon2]==1.7.4
python-multipart==0.0.6
psycopg2
anthropic==0.35.0
openai>=1.3.0,<2
httpx>=0.27.1,<1
aiohttp==3.11.7
